requirements.txt
Django==4.1.3
psycopg==3.1.4
httpx==0.23.0
pygit2==1.11.1
python-dotenv==0.21.0
Pillow==9.3.0
GitPython==3.1.29